+++ src/Passes/Blur.hpp
@@ -1,9 +1,7 @@
 #pragma once
 
-#include "fg/FrameGraphResource.hpp"
+#include "fg/Fwd.hpp"
 #include "../RenderContext.hpp"
-
-class FrameGraph;
 
 class Blur {
 public:
