+++ src/Passes/SkyboxPass.hpp
@@ -1,10 +1,7 @@
 #pragma once
 
-#include "fg/FrameGraphResource.hpp"
+#include "fg/Fwd.hpp"
 #include "../RenderContext.hpp"
-
-class FrameGraph;
-class FrameGraphBlackboard;
 
 class SkyboxPass {
 public:
