+++ src/Passes/TonemapPass.hpp
@@ -1,9 +1,7 @@
 #pragma once
 
-#include "fg/FrameGraphResource.hpp"
+#include "fg/Fwd.hpp"
 #include "../RenderContext.hpp"
-
-class FrameGraph;
 
 enum class Tonemap : uint32_t { Clamp = 0, ACES, Filmic, Reinhard, Uncharted };
