+++ src/UploadLights.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "fg/Fwd.hpp"
+#include "Light.hpp"
+#include <vector>
+
+void uploadLights(FrameGraph &, FrameGraphBlackboard &,
+                  std::vector<const Light *> &&);
