+++ src/WorldRenderer.cpp
@@ -16,6 +16,9 @@
 #include "ReflectionsData.hpp"
 #include "SceneColorData.hpp"
 
+#include "UploadFrameBlock.hpp"
+#include "UploadLights.hpp"
+
 #include "ShaderCodeBuilder.hpp"
 
 #include "LightUtility.hpp"
@@ -29,59 +32,7 @@
 
 namespace {
 
-constexpr auto kMaxNumLights = 1024u;
 constexpr auto kTileSize = 16u;
-
-__declspec(align(16)) struct GPUCamera {
-  GPUCamera(const PerspectiveCamera &camera)
-      : projection{camera.getProjection()},
-        inversedProjection{glm::inverse(projection)}, view{camera.getView()},
-        inversedView{glm::inverse(view)}, fov{camera.getFov()},
-        near{camera.getNear()}, far{camera.getFar()} {}
-
-private:
-  glm::mat4 projection{1.0f};
-  glm::mat4 inversedProjection{1.0f};
-  glm::mat4 view{1.0f};
-  glm::mat4 inversedView{1.0f};
-  float fov;
-  float near, far;
-  // Implicit padding, 4bytes
-};
-struct GPUFrameBlock {
-  float time{0.0f};
-  float deltaTime{0.0f};
-  Extent2D resolution{0};
-  GPUCamera camera;
-  uint32_t renderFeatures{0};
-};
-
-// - Position and direction in view-space
-// - Angles in radians
-__declspec(align(16)) struct GPULight {
-  GPULight(const Light &light) {
-    position = glm::vec4{light.position, light.range};
-    color = glm::vec4{light.color, light.intensity};
-    type = static_cast<uint32_t>(light.type);
-
-    if (light.type == LightType::Spot) {
-      innerConeAngle = glm::radians(light.innerConeAngle);
-      outerConeAngle = glm::radians(light.outerConeAngle);
-    }
-    if (light.type != LightType::Point) {
-      direction = glm::normalize(glm::vec4{light.direction, 0.0f});
-    }
-  }
-
-private:
-  glm::vec4 position{0.0f, 0.0f, 0.0f, 1.0f};  // .w = range
-  glm::vec4 direction{0.0f, 0.0f, 0.0f, 0.0f}; // .w = unused
-  glm::vec4 color{0.0f, 0.0f, 0.0f, 1.0f};     // .a = intensity
-  uint32_t type{0};
-  float innerConeAngle{1.0f};
-  float outerConeAngle{1.0f};
-  // Implicit padding, 4bytes
-};
 
 [[nodiscard]] bool isLightInFrustum(const Light &light,
                                     const Frustum &frustum) {
@@ -137,37 +88,6 @@ void importLightProbe(FrameGraph &fg, FrameGraphBlackboard &blackboard,
     importTexture(fg, "SpecularIBL", &lightProbe.specular);
 }
 
-void uploadLights(FrameGraph &fg, FrameGraphBlackboard &blackboard,
-                  const PerspectiveCamera &camera,
-                  std::span<const Light *> lights) {
-  blackboard.add<LightsData>() = fg.addCallbackPass<LightsData>(
-    "UploadLights",
-    [&](FrameGraph::Builder &builder, LightsData &data) {
-      constexpr auto kBufferSize =
-        sizeof(GPULight) * kMaxNumLights + sizeof(uint32_t);
-      data.buffer =
-        builder.create<FrameGraphBuffer>("LightsBuffer", {.size = kBufferSize});
-      data.buffer = builder.write(data.buffer);
-    },
-    [=](const LightsData &data, FrameGraphPassResources &resources, void *ctx) {
-      NAMED_DEBUG_MARKER("UploadLights");
-      TracyGpuZone("UploadLights");
-
-      std::vector<GPULight> gpuLights;
-      gpuLights.reserve(lights.size());
-      for (const auto *light : lights)
-        gpuLights.emplace_back(GPULight{*light});
-
-      const auto numLights = static_cast<uint32_t>(gpuLights.size());
-      constexpr auto kLightsNumberOffset = sizeof(GPULight) * kMaxNumLights;
-
-      auto &buffer = getBuffer(resources, data.buffer);
-      static_cast<RenderContext *>(ctx)
-        ->upload(buffer, 0, sizeof(GPULight) * numLights, gpuLights.data())
-        .upload(buffer, kLightsNumberOffset, sizeof(uint32_t), &numLights);
-    });
-}
-
 enum class SortOrder { FrontToBack, BackToFront };
 struct SortByDistance {
   SortByDistance(const PerspectiveCamera &camera, SortOrder order)
@@ -196,22 +116,18 @@ private:
 
 WorldRenderer::WorldRenderer(RenderContext &rc)
     : m_renderContext{rc}, m_ibl{rc}, m_transientResources{rc},
-      m_tiledLighting{rc, kMaxNumLights, kTileSize}, m_shadowRenderer{rc},
-      m_gBufferPass{rc}, m_deferredLightingPass{rc, kMaxNumLights, kTileSize},
-      m_skyboxPass{rc}, m_weightedBlendedPass{rc, kMaxNumLights, kTileSize},
-      m_transparencyCompositionPass{rc}, m_wireframePass{rc},
-      m_brightPass{rc}, m_ssao{rc}, m_ssr{rc}, m_tonemapPass{rc}, m_fxaa{rc},
-      m_gammaCorrectionPass{rc}, m_vignettePass{rc}, m_blur{rc} {
+      m_tiledLighting{rc, kTileSize}, m_shadowRenderer{rc}, m_gBufferPass{rc},
+      m_deferredLightingPass{rc, kTileSize}, m_skyboxPass{rc},
+      m_weightedBlendedPass{rc, kTileSize}, m_transparencyCompositionPass{rc},
+      m_wireframePass{rc}, m_brightPass{rc}, m_ssao{rc}, m_ssr{rc},
+      m_tonemapPass{rc}, m_fxaa{rc}, m_gammaCorrectionPass{rc},
+      m_vignettePass{rc}, m_blur{rc}, m_blit{rc}, m_finalPass{rc} {
   m_brdf = m_ibl.generateBRDF();
-  _setupPipelines();
 }
 WorldRenderer::~WorldRenderer() {
   m_renderContext.destroy(m_brdf)
     .destroy(m_globalLightProbe.diffuse)
-    .destroy(m_globalLightProbe.specular)
-
-    .destroy(m_blitPipeline)
-    .destroy(m_additiveBlitPipeline);
+    .destroy(m_globalLightProbe.specular);
 }
 
 void WorldRenderer::setSkybox(Texture &cubemap) {
@@ -238,14 +154,14 @@ void WorldRenderer::drawFrame(const RenderSettings &settings,
   blackboard.add<BRDF>().lut = importTexture(fg, "BRDF LUT", &m_brdf);
   importLightProbe(fg, blackboard, m_globalLightProbe);
 
-  // FrameBlock is persistently bound to Uniform binding at index 0
-  _uploadFrameBlock(fg, blackboard,
-                    {
-                      .deltaTime = deltaTime,
-                      .resolution = resolution,
-                      .camera = camera,
-                      .features = settings.renderFeatures,
-                    });
+  uploadFrameBlock(fg, blackboard,
+                   {
+                     .time = m_time,
+                     .deltaTime = deltaTime,
+                     .resolution = resolution,
+                     .camera = camera,
+                     .features = settings.renderFeatures,
+                   });
 
   auto visibleLights = getVisibleLights(lights, camera.getFrustum());
 
@@ -263,7 +179,7 @@ void WorldRenderer::drawFrame(const RenderSettings &settings,
   m_gBufferPass.addGeometryPass(fg, blackboard, resolution, camera,
                                 opaqueRenderables);
 
-  uploadLights(fg, blackboard, camera, visibleLights);
+  uploadLights(fg, blackboard, std::move(visibleLights));
   // Requires depth buffer, must be executed AFTER GBufferPass.
   m_tiledLighting.cullLights(fg, blackboard);
 
@@ -296,9 +212,9 @@ void WorldRenderer::drawFrame(const RenderSettings &settings,
   const auto reflections = m_ssr.addPass(fg, blackboard);
 
   if ((settings.renderFeatures & RenderFeature_SSR))
-    sceneColor.hdr = _addColor(fg, sceneColor.hdr, reflections);
+    sceneColor.hdr = m_blit.addColor(fg, sceneColor.hdr, reflections);
   if (settings.renderFeatures & RenderFeature_Bloom)
-    sceneColor.hdr = _addColor(fg, sceneColor.hdr, brightColor);
+    sceneColor.hdr = m_blit.addColor(fg, sceneColor.hdr, brightColor);
 
   sceneColor.ldr = m_tonemapPass.addPass(fg, sceneColor.hdr, settings.tonemap);
   if (settings.debugFlags & DebugFlag_Wireframe) {
@@ -319,7 +235,7 @@ void WorldRenderer::drawFrame(const RenderSettings &settings,
   if (settings.renderFeatures & RenderFeature_Vignette)
     sceneColor.ldr = afterVignette;
 
-  _present(fg, blackboard, settings.outputMode);
+  m_finalPass.compose(fg, blackboard, settings.outputMode);
   {
     ZoneScopedN("CompileFrameGraph");
     fg.compile();
@@ -342,213 +258,4 @@ void WorldRenderer::drawFrame(const RenderSettings &settings,
 
   m_time += deltaTime;
   m_transientResources.update(deltaTime);
-}
-
-void WorldRenderer::_setupPipelines() {
-  GraphicsPipeline::Builder pipelineBuilder{};
-  pipelineBuilder
-    .setDepthStencil({
-      .depthTest = false,
-      .depthWrite = false,
-    })
-    .setRasterizerState({
-      .polygonMode = PolygonMode::Fill,
-      .cullMode = CullMode::Back,
-      .scissorTest = false,
-    });
-
-  ShaderCodeBuilder shaderCodeBuilder;
-  const auto fullScreenTriangleCodeVS =
-    shaderCodeBuilder.build("FullScreenTriangle.vert");
-  const auto blitCodeFS = shaderCodeBuilder.build("Blit.frag");
-
-  auto program =
-    m_renderContext.createGraphicsProgram(fullScreenTriangleCodeVS, blitCodeFS);
-  m_blitPipeline = pipelineBuilder.setShaderProgram(program).build();
-
-  program =
-    m_renderContext.createGraphicsProgram(fullScreenTriangleCodeVS, blitCodeFS);
-  m_additiveBlitPipeline = pipelineBuilder
-                             .setBlendState(0,
-                                            {
-                                              .enabled = true,
-                                              .srcColor = BlendFactor::One,
-                                              .destColor = BlendFactor::One,
-                                            })
-                             .setShaderProgram(program)
-                             .build();
-}
-
-void WorldRenderer::_uploadFrameBlock(FrameGraph &fg,
-                                      FrameGraphBlackboard &blackboard,
-                                      const FrameInfo &frameInfo) {
-  blackboard.add<FrameData>() = fg.addCallbackPass<FrameData>(
-    "UploadFrameBlock",
-    [&](FrameGraph::Builder &builder, FrameData &data) {
-      data.frameBlock = builder.create<FrameGraphBuffer>(
-        "FrameBlock", {.size = sizeof(GPUFrameBlock)});
-      data.frameBlock = builder.write(data.frameBlock);
-
-      builder.setSideEffect();
-    },
-    [=](const FrameData &data, FrameGraphPassResources &resources, void *ctx) {
-      NAMED_DEBUG_MARKER("UploadFrameBlock");
-      TracyGpuZone("UploadFrameBlock");
-
-      const GPUFrameBlock frameBlock{
-        .time = m_time,
-        .deltaTime = frameInfo.deltaTime,
-        .resolution = frameInfo.resolution,
-        .camera = GPUCamera{frameInfo.camera},
-        .renderFeatures = frameInfo.features,
-      };
-      static_cast<RenderContext *>(ctx)->upload(
-        getBuffer(resources, data.frameBlock), 0, sizeof(GPUFrameBlock),
-        &frameBlock);
-    });
-}
-
-FrameGraphResource WorldRenderer::_addColor(FrameGraph &fg,
-                                            FrameGraphResource target,
-                                            FrameGraphResource source) {
-  fg.addCallbackPass(
-    "AddColor",
-    [&](FrameGraph::Builder &builder, auto &) {
-      builder.read(source);
-
-      target = builder.write(target);
-    },
-    [=](const auto &, FrameGraphPassResources &resources, void *ctx) {
-      NAMED_DEBUG_MARKER("AddColor");
-      TracyGpuZone("AddColor");
-
-      const auto extent =
-        resources.getDescriptor<FrameGraphTexture>(target).extent;
-      const RenderingInfo renderingInfo{
-        .area = {.extent = extent},
-        .colorAttachments = {{
-          .image = getTexture(resources, target),
-        }},
-      };
-      auto &rc = *static_cast<RenderContext *>(ctx);
-      const auto framebuffer = rc.beginRendering(renderingInfo);
-      rc.setGraphicsPipeline(m_additiveBlitPipeline)
-        .bindTexture(0, getTexture(resources, source))
-        .setUniform1ui("u_Mode", 0)
-        .drawFullScreenTriangle()
-        .endRendering(framebuffer);
-    });
-
-  return target;
-}
-void WorldRenderer::_present(FrameGraph &fg, FrameGraphBlackboard &blackboard,
-                             OutputMode outputMode) {
-  const auto [frameBlock] = blackboard.get<FrameData>();
-
-  enum Mode_ {
-    Mode_Discard = -1,
-    Mode_Default,
-    Mode_LinearDepth,
-    Mode_RedChannel,
-    Mode_GreenChannel,
-    Mode_BlueChannel,
-    Mode_AlphaChannel,
-    Mode_ViewSpaceNormals,
-  } mode{Mode_Default};
-
-  FrameGraphResource output{-1};
-  switch (outputMode) {
-    using enum OutputMode;
-
-  case Depth:
-    output = blackboard.get<GBufferData>().depth;
-    mode = Mode_LinearDepth;
-    break;
-  case Emissive:
-    output = blackboard.get<GBufferData>().emissive;
-    break;
-  case BaseColor:
-    output = blackboard.get<GBufferData>().albedo;
-    break;
-  case Normal:
-    output = blackboard.get<GBufferData>().normal;
-    // mode = Mode_ViewSpaceNormals;
-    break;
-  case Metallic:
-    output = blackboard.get<GBufferData>().metallicRoughnessAO;
-    mode = Mode_RedChannel;
-    break;
-  case Roughness:
-    output = blackboard.get<GBufferData>().metallicRoughnessAO;
-    mode = Mode_GreenChannel;
-    break;
-  case AmbientOcclusion:
-    output = blackboard.get<GBufferData>().metallicRoughnessAO;
-    mode = Mode_BlueChannel;
-    break;
-
-  case SSAO:
-    output = blackboard.get<SSAOData>().ssao;
-    mode = Mode_RedChannel;
-    break;
-  case BrightColor:
-    output = blackboard.get<BrightColorData>().brightColor;
-    break;
-  case Reflections:
-    output = blackboard.get<ReflectionsData>().reflections;
-    break;
-
-  case Accum:
-    if (auto oit = blackboard.try_get<WeightedBlendedData>(); oit)
-      output = oit->accum;
-    break;
-  case Reveal:
-    if (auto oit = blackboard.try_get<WeightedBlendedData>(); oit) {
-      output = oit->reveal;
-      mode = Mode_RedChannel;
-    }
-    break;
-
-  case LightHeatmap: {
-    const auto &lightCullingData = blackboard.get<LightCullingData>();
-    if (lightCullingData.debugMap.has_value())
-      output = *lightCullingData.debugMap;
-  } break;
-
-  case HDR:
-    output = blackboard.get<SceneColorData>().hdr;
-    break;
-  case FinalImage:
-    output = blackboard.get<SceneColorData>().ldr;
-    break;
-
-  default:
-    assert(false);
-    break;
-  }
-  if (output == -1) mode = Mode_Discard;
-
-  fg.addCallbackPass(
-    "Present",
-    [&](FrameGraph::Builder &builder, auto &) {
-      builder.read(frameBlock);
-      if (mode != Mode_Discard) builder.read(output);
-      builder.setSideEffect();
-    },
-    [=](const auto &, FrameGraphPassResources &resources, void *ctx) {
-      NAMED_DEBUG_MARKER("Present");
-      TracyGpuZone("Present");
-
-      const auto extent =
-        resources.getDescriptor<FrameGraphTexture>(output).extent;
-      auto &rc = *static_cast<RenderContext *>(ctx);
-      rc.beginRendering({.extent = extent}, glm::vec4{0.0f});
-      if (mode != Mode_Discard) {
-        rc.setGraphicsPipeline(m_blitPipeline)
-          .bindUniformBuffer(0, getBuffer(resources, frameBlock))
-          .bindTexture(0, getTexture(resources, output))
-          .setUniform1ui("u_Mode", mode)
-          .drawFullScreenTriangle();
-      }
-    });
 }
